chore: rename ChatWindow to TarotChatWindow
